requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe pyproject.toml
#
beartype==0.14.1
    # via dycw-utilities
beautifulsoup4==4.12.2
    # via parse-opml (pyproject.toml)
click==8.1.5
    # via dycw-utilities
dycw-utilities[click,more-itertools]==0.10.0
    # via parse-opml (pyproject.toml)
lxml==4.9.3
    # via parse-opml (pyproject.toml)
more-itertools==9.1.0
    # via dycw-utilities
soupsieve==2.4.1
    # via beautifulsoup4
typing-extensions==4.7.1
    # via dycw-utilities
